+++ CHANGELOG.md
@@ -1,14 +1,22 @@
 # Tideland Go REST Server Library
 
+## 2017-03-XX
+
+- Rename internal *envelope* to public *Feedback* in *rest*
+- Added *ReadFeedback()* to *Response* in *request*
+- Asserts in *restaudit* now internally increase the callstack 
+  offset so that the correct test line number is shown
+- Added *Response.AssertBodyGrep()* to *restaudit*
+
 ## 2017-02-12
 
 - Some renamings in *Request*  and *Response*, sadly
-  to the previous minor release
+  incompatible to the previous minor release
 - More convenience helpers for testing
 - Adopted new testing to more packages
 - Using http package constants instead of own
   plain strings
-- Added documentation to restaudit
+- Added documentation to *restaudit*
 
 ## 2017-02-10